--- apps/charterafrica/src/components/Grants/index.js
@@ -1,3 +1,0 @@
-import Grants from "./Grants";
-
-export default Grants;
