+++ apps/charterafrica/src/components/GrantsAndFellowshipsGrid/GrantsAndFellowshipsGrid.snap.js
@@ -1,6 +1,6 @@
 // Jest Snapshot v1, https://goo.gl/fbAQLP
 
-exports[`<Grants /> renders unchanged 1`] = `
+exports[`<GrantsAndFellowshipsGrid /> renders unchanged 1`] = `
 <div>
   <div
     class="MuiBox-root css-156ejva"
