+++ apps/charterafrica/src/components/GrantsAndFellowshipsGrid/index.js
@@ -1,0 +1,3 @@
+import GrantsAndFellowshipsGrid from "./GrantsAndFellowshipsGrid";
+
+export default GrantsAndFellowshipsGrid;
